+++ src/main/java/domain/result/ResultCalculator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
